Fixes some issues with slider values
Specifically:
- The slider would always reset to 0, even when the node was, say, 0.5.
  This seems to be fixed by forcing `Slider.start_value` to be a float.
- The slider display value would update when the simulation restarted,
  but the actual value being output would stay what it was before.
  This seems to be fixed by calling `this.notify`, as is done in the
  `user_reset_value` function.

+++ nengo_gui/components/slider.py
@@ -12,7 +12,7 @@ class Slider(Component):
         self.override = [None] * node.size_out
         self.value = np.zeros(node.size_out)
         self.label = viz.viz.get_label(node)
-        self.start_value = np.zeros(node.size_out)
+        self.start_value = np.zeros(node.size_out, dtype=float)
         self.struct = struct.Struct('<%df' % (1 + node.size_out))
         self.data = collections.deque()
         if not callable(self.base_output):
